refactor: 요일리스트 생성 방법 변경
