+++ app/src/main/AndroidManifest.xml
@@ -17,11 +17,9 @@
         <activity android:name=".MainActivity"/>
         <activity
             android:name=".SplashActivity"
-            android:exported="true"
-            >
+            android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
-
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
