+++ app/src/main/java/com/example/webtoonhub/MainActivity.kt
@@ -22,6 +22,7 @@ import androidx.lifecycle.ViewModelProvider
 import androidx.viewpager2.widget.ViewPager2
 import com.example.webtoonhub.fragment.mainFragment
 import com.example.webtoonhub.model.WebToonData
+import com.example.webtoonhub.utils.API_DAY_WEEK
 import java.util.*
 import kotlin.collections.ArrayList
 
@@ -35,24 +36,47 @@ androidx.appcompat.widget.SearchView.OnQueryTextListener {
     private lateinit var mySearchView: androidx.appcompat.widget.SearchView
     private lateinit var mySearchViewEditText: EditText
     var fragments : ArrayList<mainFragment> = ArrayList()
+
+    var dayWeek: API_DAY_WEEK = API_DAY_WEEK.MONDAY//기본요일을 월요일로 설정
     var startToDayWeeks:ArrayList<String> = ArrayList()//인텐트로 받아올 오늘요일부터 시작하는 요일 리스트
+
     var selectTab:Int = 0
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        getIntents()//오늘 요일받아오기
-
+        Log.d(Constants.TAG, "MainActivity - onCreate() called")
+        //탭 생성할 요일 리스트 생성
+        getTDayWeek()//오늘요일구하기
+        makeWeekList(dayWeek.dayNum)//오늘요일을 시작으로 요일리스트 생성
         setBinding()//바인딩
         setContentView(binding.root)
-        Log.d(Constants.TAG, "MainActivity - onCreate() called")
-
         setActionBar()// 액션바의 오버라이드 함수로 (네이버,카카오)플랫폼 선택하는 액션바 설정
         setTabsListener()//요일 선택하는 탭바의 클릭 이벤트리스너//통신
         setTabsFragment()//프래그먼트 생성하고 그수만큼 탭바 생성
         setObserve()
-        //Log.d(Constants.TAG,"MainActivity -$selectTab")
 
     }//onCreate
+
+    fun getTDayWeek(){
+        //시스템요일(일~토 1~7) 과 api요일을 판단하는 숫자가달라서 enum쿨래스로 선언하여 when을 통해 판별하여 api요일 숫자로 저장
+        val instance = Calendar.getInstance()
+        val week=instance.get(Calendar.DAY_OF_WEEK)
+        dayWeek= dayWeek.setDay(week)
+        Log.d(Constants.TAG,"MainActivity - getTDayWeek() called/dayWeek = ${dayWeek.dayWeek}//${dayWeek.dayNum}//${dayWeek.systemDayNum}")
+    }
+    //월~완결 순서를  오늘요일부터 시작하게 리스트 만들기.
+    fun makeWeekList( daynum:Int){
+        var baseDayweeks : Array<out String> = resources.getStringArray(R.array.base_week_array)//arrays.xml에서 가져오기
+        var i=daynum
+        do {
+            startToDayWeeks.add(baseDayweeks[i])
+            i++
+            if (i==7){
+                i=0
+            }
+        }while (startToDayWeeks.size<7)
+    }
+
     fun setObserve() {//
         //Log.d(Constants.TAG,"!!MainActivity - setObserve() called")
         mainViewModel.webtoonDataList.observe(this, androidx.lifecycle.Observer {
@@ -60,11 +84,7 @@ androidx.appcompat.widget.SearchView.OnQueryTextListener {
             pagerAdapter.getFragment(selectTab).setData(it)
         })
     }
-    //splashActivity로부터 인텐트 받기
-    fun getIntents() {
-        startToDayWeeks = intent.getStringArrayListExtra("startToDayWeeks") as ArrayList<String>
-        Log.d(Constants.TAG, "MainActivity - getIntents() called  = ${startToDayWeeks}")
-    }
+
     //바인딩
     fun setBinding() {
         binding = ActivityMainBinding.inflate(layoutInflater)
